+++ NEWS.md
@@ -1,5 +1,11 @@
+# brat 0.0.2.9001
+
+* Documentation written for `brainy()`.
+* The brainy app will process an existing tokenization and map character offset positions of annotations on this annotation.
+
 # brat 0.0.1.9003
 
 * New sample data `example_doc_data` and `example_coll_data`.
 * New sample file edokelley.ann in inst/sample_data dir.
 * New function `parse_annfile'
+